\name{body2string}
\alias{body2string}
\title{Converts the body of a function to a string}
\description{
Converts the body of a given function to a string allowing further manipulations of this function.
}
\usage{body2string(func)}
\arguments{
  \item{func}{A function.}
}
\details{
This is a required function of the package \kbd{ComplexAnalysis}.
}
\value{
A string (or a vector of strings if the body of the function consists of multiple lines).
}
\author{
Char Leung
}

