\name{ComplexAnalysis-package}
\alias{ComplexAnalysis-package}
\alias{ComplexAnalysis}
\docType{package}
\title{
\packageTitle{ComplexAnalysis}
}
\description{
\packageDescription{ComplexAnalysis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ComplexAnalysis}
\packageIndices{ComplexAnalysis}
 Evaluate complex integrals along a disc or a straight line. 
 Evaluate higher-order derivatives by employing the Cauchy integral formula.
 Evaluate residues.
 Search for simple poles given a function.
 Evaluate limits.
}
\author{
\packageAuthor{ComplexAnalysis}

Maintainer: \packageMaintainer{ComplexAnalysis}
}

